+++ app/src/main/java/nextstep/jwp/model/http/Param.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.model.http;
 
 public class Param {
