+++ app/src/main/java/nextstep/jwp/model/http/httpresponse/HttpResponse.java
@@ -1,4 +1,6 @@
-package nextstep.jwp.model;
+package nextstep.jwp.model.http.httpresponse;
+
+import nextstep.jwp.model.http.HttpHeaders;
 
 public class HttpResponse {
