+++ app/src/main/java/nextstep/jwp/model/http/httpresponse/HttpStatusCode.java
@@ -1,4 +1,4 @@
-package nextstep.jwp.model;
+package nextstep.jwp.model.http.httpresponse;
 
 import java.util.Arrays;
 import nextstep.jwp.exception.StatusCodeNotFoundException;
