+++ app/src/test/java/nextstep/jwp/controller/HttpControllerTest.java
@@ -1,10 +1,11 @@
 package nextstep.jwp.model;
 
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
-import static org.junit.jupiter.api.Assertions.*;
 
 import java.io.IOException;
 import nextstep.jwp.exception.DuplicatedAccountException;
+import nextstep.jwp.model.httprequest.HttpRequest;
+import nextstep.jwp.model.httpresponse.HttpResponse;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
