+++ app/src/test/java/nextstep/jwp/model/http/httprequest/HttpRequestTest.java
@@ -1,10 +1,9 @@
 package nextstep.jwp.model;
 
-import static org.junit.jupiter.api.Assertions.*;
-
 import java.io.ByteArrayInputStream;
 import java.io.IOException;
 import java.io.InputStream;
+import nextstep.jwp.model.httprequest.HttpRequest;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
