+++ app/src/test/java/nextstep/jwp/model/http/httpresponse/HttpResponseTest.java
@@ -1,7 +1,6 @@
 package nextstep.jwp.model;
 
-import static org.junit.jupiter.api.Assertions.*;
-
+import nextstep.jwp.model.httpresponse.HttpResponse;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
